Updates to obtains facility data from the CAM
Numerous updates to harmonize with the release of the CAMPD API. Namely, downloads of the facility data files are now automated. The Python scrips collects one file for each year and the R scripts combine them. The local config file has changed to include a space for an API key to the CAMPD API.